new:簡易的なメモリアロケータを実装

+++ kernel/types.h
@@ -6,8 +6,8 @@ typedef unsigned short uint16_t;
 typedef unsigned char  uint8_t;
 
 typedef uint32_t size_t;
-typedef uint32_t paddr_t;
-typedef uint32_t vaddr_t;
+typedef uint64_t paddr_t;
+typedef uint64_t vaddr_t;
 
 #define true 1
 #define false 0
